refactor(l10n): Remove hardcoded localizations and simplify demo setup
This commit removes the explicit declaration and inclusion of `AppLocalizations` and its delegates from `main.dart` and `lib/rhizosphere.dart`. It also deletes all generated ARB and Dart files related to `AppLocalizations` as they are no longer managed by the core library. This simplifies the demo setup and aligns with a more conventional approach to internationalization management.